dashboard navbarok leprogramozása és a kijelentkezés modal felvázolása

+++ Telemedicina/frontend/src/app/view/doctor/components/doctor-navbar/doctor-navbar.component.ts
@@ -1,7 +1,8 @@
 import { Component } from '@angular/core';
-import { IonicModule } from '@ionic/angular';
-import { NgIf } from '@angular/common';
+import {IonicModule, ModalController} from '@ionic/angular';
+import {NgClass, NgForOf, NgIf} from '@angular/common';
 import { Router, RouterLinkActive, RouterModule } from '@angular/router';
+import {LogoutModalComponent} from '../../../../shared/logout-modal/logout-modal.component';
 
 @Component({
   selector: 'app-doctor-navbar',
@@ -9,7 +10,9 @@ import { Router, RouterLinkActive, RouterModule } from '@angular/router';
     IonicModule,
     NgIf,
     RouterLinkActive,
-    RouterModule
+    RouterModule,
+    NgClass,
+    NgForOf
   ],
   templateUrl: './doctor-navbar.component.html',
   standalone: true,
@@ -17,23 +20,44 @@ import { Router, RouterLinkActive, RouterModule } from '@angular/router';
 })
 export class DoctorNavbarComponent {
   user: any;
-  isExpanded = false;
-  showLogoutPopup = false;
+  isCollapsed = true;
 
-  constructor(private router: Router) {}
+  constructor(private router: Router, private modalCtrl: ModalController) {}
+
+  menuItems = [
+    { icon: 'home', label: 'Profil', route: '/dashboard/doctor' },
+    { icon: 'clipboard', label: 'Új diagnózis', route: '/naplo' },
+    { icon: 'document-attach', label: 'Dokumentum feltöltés', route: '/orvos-kereso' },
+    { icon: 'calendar', label: 'Rendelési időpontjaim', route: '/idopontok' },
+    { icon: 'notifications', label: 'Értesítések', route: '/ertesitesek' },
+    { icon: 'settings-sharp', label: 'Beállítások', route: '/beallitasok' },
+    { icon: 'log-out-outline', label: 'Kijelentkezés', route: '/logout' }
+  ];
 
   toggleSidebar() {
-    this.isExpanded = !this.isExpanded;
+    this.isCollapsed = !this.isCollapsed;
   }
 
-  toggleLogoutPopup() {
-    this.showLogoutPopup = !this.showLogoutPopup;
+  async confirmLogout() {
+    const modal = await this.modalCtrl.create({
+      component: LogoutModalComponent,
+      cssClass: 'custom-logout-modal',
+    });
+
+    await modal.present();
+
+    const { data } = await modal.onDidDismiss();
+    if (data === true) {
+      localStorage.clear();
+      void this.router.navigate(['/regist-login'], { queryParams: { tab: 'login' } });
+    }
   }
 
-  confirmLogout() {
-    localStorage.clear();
-    this.user = null;
-    this.showLogoutPopup = false;
-    this.router.navigate(['/regist-login'], { queryParams: { tab: 'login' } });
+  navigateTo(route: string) {
+    void this.router.navigate([route]);
+  }
+
+  isActiveRoute(route: string): boolean {
+    return location.pathname === route;
   }
 }
